move utils to appointment utils
