+++ src/applications/vass/utils/test-utils.jsx
@@ -151,70 +151,6 @@ export const LocationDisplay = () => {
 export const TestComponent = () => (
   <div data-testid="test-component">Test Content</div>
 );
-/**
- * Default mock appointment data for tests.
- * Use createMockAppointmentData() to override specific fields.
- */
-export const defaultAppointmentData = {
-  appointmentId: 'test-appointment-123',
-  startUTC: '2025-05-01T16:00:00.000Z',
-  endUTC: '2025-05-01T16:30:00.000Z',
-  agentId: '353dd0fc-335b-ef11-bfe3-001dd80a9f48',
-  agentNickname: 'Bill Brasky',
-  appointmentStatusCode: 1,
-  appointmentStatus: 'Confirmed',
-  cohortStartUtc: '2025-01-01T00:00:00.000Z',
-  cohortEndUtc: '2025-12-31T23:59:59.999Z',
-};
-
-/**
- * Creates mock appointment data with optional overrides.
- *
- * @param {Object} overrides - Fields to override in the default appointment data
- * @returns {Object} Appointment data object
- *
- * @example
- * const appointment = createMockAppointmentData({ appointmentId: 'custom-id' });
- */
-export const createMockAppointmentData = (overrides = {}) => ({
-  ...defaultAppointmentData,
-  ...overrides,
-});
-
-/**
- * Creates RTK Query cache state with a pre-populated appointment.
- * Use this to mock the getAppointment query result.
- *
- * @param {string} appointmentId - The appointment ID to use in the cache key
- * @param {Object} appointmentData - The appointment data to cache (use createMockAppointmentData)
- * @returns {Object} RTK Query state object for vassApi
- *
- * @example
- * const vassApiState = createVassApiStateWithAppointment('123', createMockAppointmentData({ appointmentId: '123' }));
- * const options = getDefaultRenderOptions({}, { vassApi: vassApiState });
- */
-export const createVassApiStateWithAppointment = (
-  appointmentId,
-  appointmentData,
-) => ({
-  queries: {
-    [`getAppointment({"appointmentId":"${appointmentId}"})`]: {
-      status: 'fulfilled',
-      endpointName: 'getAppointment',
-      requestId: 'test',
-      startedTimeStamp: 0,
-      data: appointmentData,
-    },
-  },
-  mutations: {},
-  provided: {},
-  subscriptions: {},
-  config: {
-    online: true,
-    focused: true,
-    middlewareRegistered: true,
-  },
-});
 
 // Re-export commonly used items for convenience
 export { reducers, vassApi };
